Added screenshot, webcam, and app system logging

+++ src/ParentalSight.Core/Contracts/IKeyloggerOptions.cs
@@ -2,7 +2,7 @@
 {
     public interface IKeyloggerOptions
     {
-        long LogRotationInMilliseconds { get; }
+        int LogRotationInMilliseconds { get; }
         string OutputPath { get; }
     }
 }
